+++ server/src/http/server.ts
@@ -21,7 +21,6 @@ import eligibleTrainingsForAppointmentRoute from "./controllers/admin/eligibleTr
 import adminEtablissementRoute from "./controllers/admin/etablissement.controller"
 import formationsRoute from "./controllers/admin/formations.controller"
 import application from "./controllers/application.controller"
-import applicationRouteV2 from "./controllers/application.controller.v2"
 import appointmentRequestRoute from "./controllers/appointmentRequest.controller"
 import appointmentRequestRouteV2 from "./controllers/appointmentRequest.controller.v2"
 import { coreRoutes } from "./controllers/core.controller"
@@ -47,6 +46,7 @@ import trainingLinks from "./controllers/trainingLinks.controller"
 import unsubscribeLbaCompany from "./controllers/unsubscribeRecruteurLba.controller"
 import updateLbaCompany from "./controllers/updateRecruteurLba.controller"
 import userRoute from "./controllers/user.controller"
+import applicationRouteV2 from "./controllers/v2/application.controller.v2"
 import jobsRouteV2 from "./controllers/v2/jobs.controller.v2"
 import version from "./controllers/version.controller"
 import { auth } from "./middlewares/authMiddleware"
